requirements.txt
asgiref==3.6.0
Django==4.1.7
psycopg2-binary==2.9.5
sqlparse==0.4.3
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
django-filter==22.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
drf-yasg-json-api==0.9.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
psycopg2==2.9.5
PyJWT==2.6.0
pytz==2022.7.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
uritemplate==4.1.1
urllib3==1.26.14
